requirements.txt
# Google API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Environment management
python-dotenv>=1.0.0

# TUI
textual>=0.40.0
rich>=13.6.0

# Date handling
python-dateutil>=2.8.2
